requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
fpdf2>=2.7.0
PyYAML>=6.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
tqdm>=4.66.0
